np.int64 array type cat visits

+++ algorithms/lib/cat_utils.py
@@ -67,7 +67,7 @@ def get_users_cat_visits(training_matrix,poi_cats):
         for lid in lids:
             for cat in poi_cats[lid]:
                 cats_visits[cat]+=training_matrix[i,lid]
-        cv=np.array(list(cats_visits.values()))
+        cv=np.array(list(cats_visits.values()),dtype=np.int64)
         users_cv.append(cv)
     return users_cv
